Rename decorator module, bump dev version
