+++ cli_test_helpers/__init__.py
@@ -5,14 +5,14 @@ __author__ = 'Peter Bittner'
 __email__ = 'peter@painless.software'
 __license__ = 'GPLv3'
 __url__ = 'https://github.com/painless-software/python-cli-test-helpers'
-__version__ = '1.0.1'
+__version__ = '2.0.0-dev0'
 
 __all__ = [
     'ArgvContext',
     'EnvironContext',
 ]
 
-from .decorator import (
+from .decorators import (
     ArgvContext,
     EnvironContext,
 )
